+++ examples/kafka_connect/provider.tf
@@ -2,13 +2,12 @@ terraform {
   required_providers {
     aiven = {
       source  = "aiven/aiven"
-      version = ">= 2.0.0, < 3.0.0"
+      version = ">=3.0.0"
     }
   }
 }
 
 # Initialize provider. No other config options than api_token
 provider "aiven" {
-  api_token = var.aiven_api_token
+  api_token = var.avn_token
 }
-